Added support for json content in ecv_map attribute, lb component.
The value part of ecv_map hash can now use json content, which may be an array
of health probes, or a hash of a single health probe. This way we can specify
non-default values for a hash probe.
Also we now can use a health probe with a different protocol (and port) as
compared to one the LB rule has. Unit and integration tests have been adjusted
to reflect the change in functionality.
Multiple health probes can be specified per LB rule but only the one that has
default attribute which equals true, will be associated with the LB rule.