Fixed a bug where a failed deployment puts system into an unrecoverable state.
When volume recipe fails at attach step (for example due to a network blip),
it cannot gracefully recover after that failure. The reason is it's using
service files for each attached volume to store information about what device
the volume got attachced to. In case of failure that file gets no data and that
messes up all the following logic as it's heavily relied on the device id of
the attached volume.
To fix that I added extra logic to query provider data to determine the device
id.
Also I synced up the latest changes from main-1 circuit, to place ephemeral
volumes in azure onto root.